pyproject.toml
[project]
name = "my-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "sidra1314", email = "sidrakhalid1314@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "openai-agents>=0.0.14",
    "python-dotenv>=1.1.0",
]

[project.scripts]
my-agent = "my_agent:main"
run-global = "my_agent.global_config:run_global"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
